fix: Explicity disconnect expression Prisma client
By default, a Prisma client will only disconnect when the Node.js process ends.
When we generate RLS expressions, we only need to keep the client
available for the duration of the expression calculation and it can be
disconnected and disposed of afterwards.

See https://www.prisma.io/docs/concepts/components/prisma-client/working-with-prismaclient/connection-management

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>