fix: BAC-129 increase base ITX timeout to 30s
This change increases the timeout of the base interactive transaction
the Yates client uses for queries to 30 seconds, up from the default
value of 5s. This is a lot less agressive and will better tolerate
large/slow queries.
Ideally this value should be configurable when generating the client,
but I will leave this for the future.

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>