fix: use upstream node-sql-parser package
This changes Yates to use the original upstream node-sql-parser package.
The only reason Yates uses my fork is because at the time the
`exprToSQL` method I contributed was not available in the general
release. Fast-forward to 10 months later and my fork is now 499 commits
behind master XD
The only real difference that has to be made is the way that `$n`
parameters are represented in the AST, but this is a simple change to
make.

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>