feat: use an ability table for handling ability updates
This change adds a new table, `_yates_abilities` that is used to track
abilities in the DB. Yates will check against this table to see if an
ability needs to be created or updated, and will write to the table
after the policy has been modified. This allows us to check strict
equivalence of expressions, something which can't be done on the
`pg_catalog.pg_policies` table due to Postgres performing formatting on
the expression when we create or update the policy, resuling in an
expression that is very hard to check for equality with the expression
coming from userland.
This table will be created and populated automatically if it doesn't
already exist.