fix: avoid race condition when creating policies
If you're trying to startup multiple Yates clients simultaneously, it's
possible that you hit an error where on client tries to create a policy
that the other client has already created.
This occurs because we read *all* abilities in one go. To avoid this, we
check the ability table once per RLS policy, and we do it inside
a transaction with a lock. This means that the second client will behave
correctly and not try to create the policy again.