fix: remove edge case chance of SET ROLE leaking
This changes removes the edge case chance that `SET ROLE` is never reset
either by a rollback or `SET ROLE none`, by switching to `SET LOCAL
ROLE` which will last only until the end of the current transaction.

Whilst the scenario where the role leaks outside of the transaction
without being reset correctly is extremely unlikely, and I am unable to
create test scenario that replicates it, I think it is wise to future
proof the code and make further guarantees that the leak cannot occur.

See:
- https://www.postgresql.org/docs/14/sql-set-role.html
- https://www.postgresql.org/docs/14/sql-set.html